Starts refactoring
We want to break out the crucial functions into the `nytid.courses`
module and keep only the CLI functions in `nytid.cli.courses`. So that
`nytid.cli.courses` will use `nytid.courses`.

In this work we also want to rename courses directories to course
registers. And have a `registry` module managing all registers.